pyproject.toml
[project]
name = "financial-advisor-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.129.0",
    "google-generativeai>=0.8.6",
    "inngest>=0.5.15",
    "langchain>=1.2.10",
    "langchain-google-genai>=4.2.1",
    "langgraph>=1.0.9",
    "matplotlib>=3.10.8",
    "numpy>=2.4.1",
    "pandas>=3.0.0",
    "pennylane>=0.44.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.2.2",
    "torch>=2.10.0",
    "transformers>=5.0.0",
    "uvicorn>=0.41.0",
    "yfinance>=1.1.0",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "ipykernel>=7.1.0",
    "jupyterlab>=4.5.3",
    "pytest>=9.0.2",
    "ruff>=0.14.14",
]
